test recongifuration
-reconfigures tests to come into line with CI standards
