+++ q2_quality_control/_stats.py
@@ -9,6 +9,7 @@
 from qiime2.plugin import SemanticType, model
 from q2_types.feature_data import FeatureData
 
+
 # defines types for DecontamScore
 DecontamScore = SemanticType('DecontamScore',
                              variant_of=FeatureData.field['type'])
