+++ q2_quality_control/tests/test_quality_control.py
@@ -5,6 +5,7 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
+import unittest
 import numpy as np
 import numpy.testing as npt
 import pandas as pd
