+++ setup.py
@@ -24,8 +24,7 @@ setup(
     },
     package_data={
         'q2_quality_control': ['citations.bib', 'assets/*'],
-        'q2_quality_control.tests': ['data/*'],
-        'q2_quality_control.tests.data': ['expected/*'],
+        'q2_quality_control.tests': ['data/*', 'data/expected/*'],
         'q2_quality_control._threshold_graph': ['assets/index.html'],
     },
     zip_safe=False,
